Fix bug rendering documents in file tree
